+++ SchoolDatabase/Listeners/Tables/Students_Scholarships.sql
@@ -3,7 +3,7 @@
 	StudentScholarshipId int IDENTITY (1,1),
 	StudentId int NOT NULL,
 	ScholarshipId int NOT NULL,
-	StudentAvg float,
+	StudentAvg decimal(3,2),
 	CONSTRAINT FK_Students_ScholarshipsStudents FOREIGN KEY (StudentId) REFERENCES listeners.Students ON DELETE CASCADE,
 	CONSTRAINT FK_Students_ScholarshipsScholarships FOREIGN KEY (ScholarshipId) REFERENCES utilities.Scholarships,
 	CONSTRAINT PK_StudentScholarship PRIMARY KEY (StudentScholarshipId)
